Inventory Size fix and progression for downgrade mechanic.
Fixed an issue where all Container Tiers had the same Inventory Size due to forgetting to multiply the Inventory size by the Containers Tier.
Finished most of the implementation for downgrading, only missing the last step to assign the downgraded Containers to the new tables (and probably also assigning a new ID since the prototype changes).
Now the only missing element is to handle Container Deaths.